refactor: refactor the data structure of dataset and corresponding code in insight card
