+++ src/ControlBar.tsx
@@ -1,22 +1,32 @@
 import React from 'react'
-import ControlFont from './components/ControlFont'
-import ControlSelectedCards from './components/ControlSelectedCards'
-import ControlBigGraphSparkLine from './components/ControlBigGraphSparkLine'
-import ControlGlobalBoolean from './components/ControlGlobalBoolean'
-import ControlSelectedData from './components/ControlSelectedData'
-import ControlLineHeight from './components/ControlLineHeight'
-import ControlAspectRatio from './components/ControlAspectRatio'
-import ControlSparkLinePos from './components/ControlSparkLinePos'
+import { Col, Row } from 'antd'
+import ControlFont from './components/ControlGlobalSetting/ControlFont'
+import ControlGlobalBoolean from './components/ControlGlobalSetting/ControlGlobalBoolean'
+import ControlLineHeight from './components/ControlGlobalSetting/ControlLineHeight'
+import ControlAspectRatio from './components/ControlGlobalSetting/ControlAspectRatio'
+import ControlSparkLinePos from './components/ControlGlobalSetting/ControlSparkLinePos'
 
 const ControlBar: React.FC = () => (
   <div>
-    <ControlLineHeight />
-    <ControlSparkLinePos />
-    {/* <ControlAspectRatio /> */}
-    {/* <ControlBigGraphSparkLine /> */}
-    {/* <ControlFont />
-    <ControlSelectedCards />
-    <ControlSelectedData /> */}
+    <Row gutter={16}>
+      <Col span={8}>
+        <ControlGlobalBoolean />
+      </Col>
+      <Col span={8}>
+        <ControlFont />
+      </Col>
+      <Col span={8}>
+        <ControlLineHeight />
+      </Col>
+    </Row>
+    <Row gutter={16}>
+      <Col span={8}>
+        <ControlSparkLinePos />
+      </Col>
+      <Col span={8}>
+        <ControlAspectRatio />
+      </Col>
+    </Row>
   </div>
 )
