+++ src/components/InsightCards.tsx
@@ -5,224 +5,22 @@ import { HTML5Backend } from 'react-dnd-html5-backend'
 import { useSelector } from 'react-redux'
 import { InsightCard } from './InsightCard'
 import { AppState } from '../store'
+import { Card } from '../types'
 
 const { Header, Footer, Sider, Content } = Layout
-type Phrase = {
-  type: string
-  value: string
-  metadata?: any
-}
-type Card = {
-  key: string
-  type: string
-  BigChartData: number[]
-  phrases: Phrase[]
-}
-interface InsightCardProps {
-  type: string
-  BigChartData: number[]
-  phrasesLists: Phrase[][]
+
+interface InsightCardProps extends Card {
   id: string
+  onDrop: (id: string, targetId: string) => void
 }
-// const CardsDataset: Card[] = [
-//   {
-//     key: 'Card1',
-//     type: 'PieChart',
-//     BigChartData: [2, 4, 55, 77, 99],
-//     phrases: [
-//       { type: 'entity', value: 'DAU', metadata: { entityType: 'metric_name' } },
-//       { type: 'text', value: ' ' },
-//       {
-//         type: 'entity',
-//         value: '1.23亿',
-//         metadata: { entityType: 'metric_value', origin: 123077.34 },
-//       },
-//       { type: 'text', value: '，环比昨日 ' },
-//       {
-//         type: 'entity',
-//         value: '80万',
-//         metadata: { entityType: 'delta_value', assessment: 'positive' },
-//       },
-//       { type: 'text', value: '（' },
-//       {
-//         type: 'entity',
-//         value: '2.3%',
-//         metadata: { entityType: 'ratio_value', assessment: 'positive' },
-//       },
-//       { type: 'text', value: '）。' },
-//       { type: 'text', value: '最近 3 个动态 7 天' },
-//       {
-//         type: 'entity',
-//         value: '趋势上涨',
-//         metadata: {
-//           entityType: 'trend_desc',
-//           detail: [1, 2, 6, 18, 24, 48],
-//         },
-//       },
-//       { type: 'text', value: '。' },
-//       { type: 'text', value: '按垂直行业分：' },
-//     ],
-//   },
-//   {
-//     key: 'Card2',
-//     type: 'PieChart',
-//     BigChartData: [2, 4, 55, 7, 99],
-//     phrases: [
-//       { type: 'entity', value: 'DAU', metadata: { entityType: 'metric_name' } },
-//       { type: 'text', value: ' ' },
-//       {
-//         type: 'entity',
-//         value: '22.23亿',
-//         metadata: { entityType: 'metric_value', origin: 123077.34 },
-//       },
-//       { type: 'text', value: '，环比昨日 ' },
-//       {
-//         type: 'entity',
-//         value: '80万',
-//         metadata: { entityType: 'delta_value', assessment: 'positive' },
-//       },
-//       { type: 'text', value: '（' },
-//       {
-//         type: 'entity',
-//         value: '2.3%',
-//         metadata: { entityType: 'ratio_value', assessment: 'positive' },
-//       },
-//       { type: 'text', value: '）。' },
-//       { type: 'text', value: '最近 3 个动态 7 天' },
-//       {
-//         type: 'entity',
-//         value: '趋势上涨',
-//         metadata: {
-//           entityType: 'trend_desc',
-//           detail: [1, 2, 6, 18, 24, 48],
-//         },
-//       },
-//       { type: 'text', value: '。' },
-//       { type: 'text', value: '按垂直行业分：' },
-//     ],
-//   },
-//   {
-//     key: 'Card3',
-//     type: 'PieChart',
-//     BigChartData: [2, 2, 2, 2, 99],
-//     phrases: [
-//       { type: 'entity', value: 'DAU', metadata: { entityType: 'metric_name' } },
-//       { type: 'text', value: ' ' },
-//       {
-//         type: 'entity',
-//         value: '22.23亿',
-//         metadata: { entityType: 'metric_value', origin: 123077.34 },
-//       },
-//       { type: 'text', value: '，环比昨日 ' },
-//       {
-//         type: 'entity',
-//         value: '80万',
-//         metadata: { entityType: 'delta_value', assessment: 'positive' },
-//       },
-//       { type: 'text', value: '（' },
-//       {
-//         type: 'entity',
-//         value: '2.3%',
-//         metadata: { entityType: 'ratio_value', assessment: 'positive' },
-//       },
-//       { type: 'text', value: '）。' },
-//       { type: 'text', value: '最近 3 个动态 7 天' },
-//       {
-//         type: 'entity',
-//         value: '趋势上涨',
-//         metadata: {
-//           entityType: 'trend_desc',
-//           detail: [1, 2, 6, 18, 24, 48],
-//         },
-//       },
-//       { type: 'text', value: '。' },
-//       { type: 'text', value: '按垂直行业分：' },
-//     ],
-//   },
-//   {
-//     key: 'Card4',
-//     type: 'PieChart',
-//     BigChartData: [333, 4, 55, 7, 99],
-//     phrases: [
-//       { type: 'entity', value: 'DAU', metadata: { entityType: 'metric_name' } },
-//       { type: 'text', value: ' ' },
-//       {
-//         type: 'entity',
-//         value: '22.23亿',
-//         metadata: { entityType: 'metric_value', origin: 123077.34 },
-//       },
-//       { type: 'text', value: '，环比昨日 ' },
-//       {
-//         type: 'entity',
-//         value: '80万',
-//         metadata: { entityType: 'delta_value', assessment: 'positive' },
-//       },
-//       { type: 'text', value: '（' },
-//       {
-//         type: 'entity',
-//         value: '2.3%',
-//         metadata: { entityType: 'ratio_value', assessment: 'positive' },
-//       },
-//       { type: 'text', value: '）。' },
-//       { type: 'text', value: '最近 3 个动态 7 天' },
-//       {
-//         type: 'entity',
-//         value: '趋势上涨',
-//         metadata: {
-//           entityType: 'trend_desc',
-//           detail: [1, 2, 6, 18, 24, 48],
-//         },
-//       },
-//       { type: 'text', value: '。' },
-//       { type: 'text', value: '按垂直行业分：' },
-//     ],
-//   },
-//   {
-//     key: 'Card5',
-//     type: 'PieChart',
-//     BigChartData: [2, 222, 55, 222, 99],
-//     phrases: [
-//       { type: 'entity', value: 'DAU', metadata: { entityType: 'metric_name' } },
-//       { type: 'text', value: ' ' },
-//       {
-//         type: 'entity',
-//         value: '22.23亿',
-//         metadata: { entityType: 'metric_value', origin: 123077.34 },
-//       },
-//       { type: 'text', value: '，环比昨日 ' },
-//       {
-//         type: 'entity',
-//         value: '80万',
-//         metadata: { entityType: 'delta_value', assessment: 'positive' },
-//       },
-//       { type: 'text', value: '（' },
-//       {
-//         type: 'entity',
-//         value: '2.3%',
-//         metadata: { entityType: 'ratio_value', assessment: 'positive' },
-//       },
-//       { type: 'text', value: '）。' },
-//       { type: 'text', value: '最近 3 个动态 7 天' },
-//       {
-//         type: 'entity',
-//         value: '趋势上涨',
-//         metadata: {
-//           entityType: 'trend_desc',
-//           detail: [1, 2, 6, 18, 24, 48],
-//         },
-//       },
-//       { type: 'text', value: '。' },
-//       { type: 'text', value: '按垂直行业分：' },
-//     ],
-//   },
-// ]
+
 const InsightCards: React.FC = () => {
   const { dataset, showBigGraph, showSparkLine, selectedCards } = useSelector(
     (state: AppState) => state.system,
   )
   // console.log('datasetttttttttttttt', dataset)
   const CardNum: number = dataset.length
-  const CardsId: string[] = dataset.map((card) => card.key)
+  const CardsId: string[] = dataset.map((card) => card.CardName)
   // cards是卡片id的列表，是[Card1,Card2]
   const [cards, setCards] = useState(CardsId)
 
@@ -238,7 +36,7 @@ const InsightCards: React.FC = () => {
   }
   const Cards = cards
     .map((cardId) => {
-      const card = dataset.find((d) => d.key === cardId)
+      const card = dataset.find((d) => d.CardName === cardId)
       return {
         ...card,
         id: cardId,
@@ -249,16 +47,11 @@ const InsightCards: React.FC = () => {
   return (
     <div id='ShowInsightCards'>
       <DndProvider backend={HTML5Backend}>
-        {Cards.map((curDataset) => {
-          if (!curDataset.type) {
-            throw new Error(`No data found for the date: ${curDataset.type}`)
-          }
-          return (
-            <Content key={curDataset.key}>
-              <InsightCard {...(curDataset as InsightCardProps)} onDrop={swapCards} />
-            </Content>
-          )
-        })}
+        {Cards.map((curDataset) => (
+          <Content key={curDataset.CardName}>
+            <InsightCard {...(curDataset as InsightCardProps)} onDrop={swapCards} />
+          </Content>
+        ))}
       </DndProvider>
     </div>
   )
